+++ core/blockchain_test.go
@@ -46,6 +46,7 @@ var (
 		TriePrefetcherParallelism: 4,
 		Pruning:                   true, // Enable pruning
 		CommitInterval:            4096,
+		StateHistory:              32,
 		SnapshotLimit:             256,
 		AcceptorQueueLimit:        64,
 	}
@@ -131,6 +132,7 @@ func TestPruningBlockChainSnapsDisabled(t *testing.T) {
 				TriePrefetcherParallelism: 4,
 				Pruning:                   true, // Enable pruning
 				CommitInterval:            4096,
+				StateHistory:              32,
 				SnapshotLimit:             0, // Disable snapshots
 				AcceptorQueueLimit:        64,
 			},
@@ -181,6 +183,7 @@ func TestPruningBlockChainUngracefulShutdownSnapsDisabled(t *testing.T) {
 				TriePrefetcherParallelism: 4,
 				Pruning:                   true, // Enable pruning
 				CommitInterval:            4096,
+				StateHistory:              32,
 				SnapshotLimit:             0, // Disable snapshots
 				AcceptorQueueLimit:        64,
 			},
@@ -217,6 +220,7 @@ func TestEnableSnapshots(t *testing.T) {
 				TriePrefetcherParallelism: 4,
 				Pruning:                   true, // Enable pruning
 				CommitInterval:            4096,
+				StateHistory:              32,
 				SnapshotLimit:             snapLimit,
 				AcceptorQueueLimit:        64,
 			},
@@ -412,6 +416,7 @@ func TestUngracefulAsyncShutdown(t *testing.T) {
 				TriePrefetcherParallelism: 4,
 				Pruning:                   true,
 				CommitInterval:            4096,
+				StateHistory:              32,
 				SnapshotLimit:             256,
 				SnapshotNoBuild:           true, // Ensure the test errors if snapshot initialization fails
 				AcceptorQueueLimit:        1000, // ensure channel doesn't block
@@ -673,6 +678,7 @@ func TestTxLookupBlockChain(t *testing.T) {
 		Pruning:                   true,
 		CommitInterval:            4096,
 		SnapshotLimit:             256,
+		StateHistory:              32,
 		SnapshotNoBuild:           true, // Ensure the test errors if snapshot initialization fails
 		AcceptorQueueLimit:        64,   // ensure channel doesn't block
 		TransactionHistory:        5,
@@ -695,6 +701,7 @@ func TestTxLookupSkipIndexingBlockChain(t *testing.T) {
 		TriePrefetcherParallelism: 4,
 		Pruning:                   true,
 		CommitInterval:            4096,
+		StateHistory:              32,
 		SnapshotLimit:             256,
 		SnapshotNoBuild:           true, // Ensure the test errors if snapshot initialization fails
 		AcceptorQueueLimit:        64,   // ensure channel doesn't block