+++ eth/ethconfig/config.go
@@ -140,7 +140,7 @@ type Config struct {
 
 	// HistoricalProofQueryWindow is the number of blocks before the last accepted block to be accepted for state queries.
 	// For archive nodes, it defaults to 43200 and can be set to 0 to indicate to accept any block query.
-	// For non-archive nodes, it is forcibly set to the value of [core.TipBufferSize].
+	// For non-archive nodes, it is forcibly set to the value of StateHistory.
 	HistoricalProofQueryWindow uint64
 
 	// AllowUnprotectedTxs allow unprotected transactions to be locally issued.